Move code utilities from `stateful_transforms` to their own module, and add the ability to specify the aliases templates.
