+++ formulaic/utils/stateful_transforms.py
@@ -1,8 +1,6 @@
 import ast
 import functools
 import inspect
-import keyword
-import re
 from typing import (
     Any,
     Callable,
@@ -16,9 +14,8 @@ from typing import (
 )
 
 import astor
-import numpy
 
-from .iterators import peekable_iter
+from .code import sanitize_variable_names
 from .layered_mapping import LayeredMapping
 from .variables import get_expression_variables, Variable
 
@@ -229,93 +226,3 @@ def _is_stateful_transform(node: ast.AST, env: Mapping) -> bool:
         return getattr(func, "__is_stateful_transform__", False)
     except NameError:
         return False
-
-
-# Variable sanitization
-
-
-UNQUOTED_BACKTICK_MATCHER = re.compile(
-    r"(\\\"|\"(?:\\\"|[^\"])*\"|\\'|'(?:\\'|[^'])*'|`)"
-)
-
-
-def sanitize_variable_names(
-    expr: str, env: MutableMapping, aliases: MutableMapping
-) -> str:
-    """
-    Sanitize any variables names in the expression that are not valid Python
-    identifiers and are surrounded by backticks (`). This allows use of field
-    names that are not valid Python names.
-
-    This function transforms `expr` into a new expression where identifiers that
-    would cause `SyntaxError`s are transformed into valid Python identifiers.
-    E.g. "func(`1a`)" -> "func(_1a)". `env` is updated to reflect the mapping of
-    the old identifier to the new one, provided that the original variable name
-    was already present.
-
-    Args:
-        expr: The expression to sanitize.
-        env: The environment to keep updated with any name substitutions. This
-            environment mapping will be mutated in place during this evaluation.
-        aliases: A dictionary/map to update with any variable mappings. Mapping
-            is from the sanitized variable back to the original variable.
-
-    Returns:
-        The sanitized expression.
-    """
-
-    expr_parts = peekable_iter(UNQUOTED_BACKTICK_MATCHER.split(expr))
-
-    sanitized_expr = []
-
-    for expr_part in expr_parts:
-        if expr_part == "`":
-            variable_name_parts = []
-            while expr_parts.peek(None) not in ("`", None):
-                variable_name_parts.append(next(expr_parts))
-            variable_name = "".join(variable_name_parts)
-            if expr_parts.peek(None) is None:
-                sanitized_expr.append(f"`{variable_name}")
-            else:
-                next(expr_parts)
-                new_name = sanitize_variable_name(variable_name, env)
-                aliases[new_name] = variable_name
-                sanitized_expr.append(f" {new_name} ")
-        else:
-            sanitized_expr.append(expr_part)
-
-    return "".join(sanitized_expr).strip()
-
-
-def sanitize_variable_name(name: str, env: MutableMapping) -> str:
-    """
-    Generate a valid Python variable name for variable identifier `name`.
-
-    Args:
-        name: The variable name to sanitize.
-        env: The mapping of variable name to values in the evaluation
-            environment. If `name` is present in this mapping, an alias is
-            created for the same value for the new variable name.
-    """
-    if name.isidentifier() or keyword.iskeyword(name):
-        return name
-
-    # Compute recognisable basename
-    base_name = "".join([char if re.match(r"\w", char) else "_" for char in name])
-    if base_name[0].isdigit():
-        base_name = "_" + base_name
-
-    # Verify new name is not in env already, and if not add a random suffix.
-    new_name = base_name
-    while new_name in env:
-        new_name = (
-            base_name
-            + "_"
-            + "".join(numpy.random.choice(list("abcefghiklmnopqrstuvwxyz"), 10))
-        )
-
-    # Reuse the value for `name` for `new_name` also.
-    if name in env:
-        env[new_name] = env[name]
-
-    return new_name
